shopping-assistant-chatbot/requirements.txt
# Core Dependencies
strands-agents>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
boto3>=1.34.0
httpx>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-json-logger>=2.0.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.92.0
